[BugFix] Removed type casting from int to byte in method writeByte(int).
git-svn-id: svn://svn.code.sf.net/p/tigervnc/code/trunk@3454 3789f03b-4d11-0410-bbf8-ca57d06f2519

+++ java/src/com/tightvnc/vncviewer/RecordOutputStream.java
@@ -28,7 +28,7 @@ public class RecordOutputStream implements DataOutput {
 
   public void writeByte(int v) throws IOException {
     if (recordInterface.canWrite()) {
-      recordInterface.writeByte((byte)v);
+      recordInterface.writeByte(v);
     }
   }
